requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
blinker==1.9.0
    # via flask
certifi==2025.1.31
    # via requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via flask
colorama==0.4.6
    # via click
flask==3.1.0
    # via
    #   -r requirements.in
    #   flask-sqlalchemy
flask-sqlalchemy==3.1.1
    # via -r requirements.in
greenlet==3.1.1
    # via sqlalchemy
idna==3.10
    # via requests
itsdangerous==2.2.0
    # via flask
jinja2==3.1.5
    # via flask
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
python-dotenv==1.0.1
    # via -r requirements.in
requests==2.32.3
    # via -r requirements.in
sqlalchemy==2.0.37
    # via flask-sqlalchemy
typing-extensions==4.12.2
    # via sqlalchemy
urllib3==2.3.0
    # via requests
werkzeug==3.1.3
    # via flask
